Create Project 10.py
Write a program that asks the user for a student number and prints whether it is a valid student number or not. A student number is valid if the sum of all the digits in the student number divided by the number of digits of the student number is 0. Repeatedly ask the user if they want to enter another
student number (Yes / No). Stop the programming upon “no”
Print either “Valid” or “Invalid” as output!
If the user enters 0, print “Thank you”